+++ NEWS.md
@@ -5,6 +5,7 @@
 * add masking based on pixel band values while reading images, see `?image_mask`
 * collection format Sentinel2_L2A now includes WVP, AOT, and SCL bands 
 * add `write_json_descr`argument to `write_ncdf()`
+* new argument `with_VRT` in `write_ncdf()` to write GDAL VRT datasets for data cube time slices
 * `write_ncdf`, `as_stars()`, and `plot()` are now  interruptible, though it might take some time to let all threads finish processing their current chunk
 * new image collection database schema, existing collections must be recreated
 * add `animate()` function to create data cube time series animations
@@ -13,6 +14,8 @@
 * new collection format for Sentinel-2 data on Theia
 * add `as_array()` function to convert a data cube to a native in-memory R array
 * collection formats can now overwrite scale, offset, and unit for bands
+* add `write_COG` to export data cube time slices as cloud-optimized GeoTIFF files
+
 
 
 # gdalcubes 0.1.0 (2019-05-15)